refactor(core_agent): update builder wiring and documentation
- Updated AgentBuilder and AppBuilder with new domain-oriented import paths
- Cleaned up plugins/__init__.py and updated package README
- Synchronized entry points with the new directory structure
