+++ agent/core_agent/plugins/README.md
@@ -1,20 +1,21 @@
-# Plugins Module
+# Agent Plugins
 
-This module contains lifecycle interceptors and post-turn callbacks that orchestrate the agent's behavior.
+This module contains integrated behaviors and message interceptors that extend the ADK Application logic.
 
-## User Uploads Ingestion
+## Plugins Overview
 
-The `GeminiEnterpriseFileIngestionPlugin` is responsible for ensuring that any file uploaded by a user is correctly persisted and referenced.
+### `ingestion/`
+- **`GeminiEnterpriseFileIngestionPlugin`**: Intercepts user messages to persist inline binary data to GCS and replace it with URI references. This ensures files uploaded via Gemini Enterprise are available for the agent's multi-modal analysis.
 
-### The Lifecycle
-1.  **Interception**: The plugin hooks into `on_user_message_callback`.
-2.  **Pre-Scan**: It scans the message for existing GCS references to populate a **Turn Registry**.
-3.  **Inline Persistence**: It saves any raw `inline_data` bytes to GCS and replaces them with `file_data` references.
-4.  **Tag Resolution (Discovery)**: It searches for GE tags (e.g., `<start_of_user_uploaded_file: ...>`). If the tags are empty, it uses the **Lazy Association** pattern to find the file in GCS (pre-stashed by Gemini Enterprise) and injects the correct `file_data` part.
+## Registration
 
-## Artifact Rendering
+Plugins are registered in the `App` (or `AdkApp`) via the `AppBuilder`:
 
-The `storage/` sub-package contains the `render_pending_artifacts` callback.
+```python
+from .plugins.ingestion.plugin import GeminiEnterpriseFileIngestionPlugin
 
-- **Purpose**: Gemini Enterprise cannot render GCS URIs visually in the final response.
-- **Mechanism**: This callback (registered as an `after_agent_callback`) scans the session state for pending artifacts, downloads them as bytes from the `StorageService`, and returns them as inline parts so they appear in the Gemini UI.
+App(
+    ...
+    plugins=[GeminiEnterpriseFileIngestionPlugin()]
+)
+```
