+++ agent/core_agent/plugins/__init__.py
@@ -1,21 +1,5 @@
-from .storage import (
-    GetArtifactUriRequest,
-    GetArtifactUriResponse,
-    GetArtifactUriTool,
-    ImportGcsToArtifactRequest,
-    ImportGcsToArtifactResponse,
-    ImportGcsToArtifactTool,
-    render_pending_artifacts,
-)
-from .user_uploads import GeminiEnterpriseFileIngestionPlugin
+from .ingestion.plugin import GeminiEnterpriseFileIngestionPlugin
 
 __all__ = [
-    "render_pending_artifacts",
-    "GetArtifactUriTool",
-    "ImportGcsToArtifactTool",
-    "GetArtifactUriRequest",
-    "GetArtifactUriResponse",
-    "ImportGcsToArtifactRequest",
-    "ImportGcsToArtifactResponse",
     "GeminiEnterpriseFileIngestionPlugin",
 ]
